bắn theo phương chuẩn hơn

+++ src/models/BulletModel1.java
@@ -62,12 +62,12 @@ public class BulletModel1 extends GameModel implements GameModelCanMove {
     public void angleChangeRight(){
 
         if(angle > 0){
-            speedY = -(int) (speed*Math.sin(Math.toRadians(angle)));
-            speedX = (int) (speed*Math.cos(Math.toRadians(angle)));
+            speedY = -(float)(speed*Math.sin(Math.toRadians(angle)));
+            speedX = (float) (speed*Math.cos(Math.toRadians(angle)));
         }
         else if(angle < 0){
-            speedY = -(int) (speed*Math.sin(Math.toRadians(angle)));
-            speedX = (int) (speed*Math.cos(Math.toRadians(angle)));
+            speedY = -(float) (speed*Math.sin(Math.toRadians(angle)));
+            speedX = (float) (speed*Math.cos(Math.toRadians(angle)));
         } else if(angle == 0){
             speedY = 0;
             speedX = speed;
